LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.2_The_while_statement/S12.6.2_A6_T6-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Expected token '('
S12.6.2_A6_T6

PASS Expected parsing failure

TEST COMPLETE
